refactor: move shared utility functions to utils
* Also applies to shared util function in tests
