+++ internal/framework/datasources/certificate_data_source_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/terraform-providers/terraform-provider-dnsimple/internal/framework/provider"
-	"github.com/terraform-providers/terraform-provider-dnsimple/internal/framework/utils"
+	"github.com/terraform-providers/terraform-provider-dnsimple/internal/framework/test_utils"
 )
 
 func TestAccCertificateDataSource(t *testing.T) {
@@ -18,7 +18,7 @@ func TestAccCertificateDataSource(t *testing.T) {
 	domain := os.Getenv("DNSIMPLE_DOMAIN")
 	certificateId := os.Getenv("DNSIMPLE_CERTIFICATE_ID")
 	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { utils.TestAccPreCheck(t) },
+		PreCheck:                 func() { test_utils.TestAccPreCheck(t) },
 		ProtoV6ProviderFactories: provider.NewProto6ProviderFactory(),
 		Steps: []resource.TestStep{
 			{
