+++ internal/framework/datasources/zone_data_source_test.go
@@ -7,12 +7,12 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/terraform-providers/terraform-provider-dnsimple/internal/framework/provider"
-	"github.com/terraform-providers/terraform-provider-dnsimple/internal/framework/utils"
+	"github.com/terraform-providers/terraform-provider-dnsimple/internal/framework/test_utils"
 )
 
 func TestAccZoneDataSource(t *testing.T) {
 	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { utils.TestAccPreCheck(t) },
+		PreCheck:                 func() { test_utils.TestAccPreCheck(t) },
 		ProtoV6ProviderFactories: provider.NewProto6ProviderFactory(),
 		Steps: []resource.TestStep{
 			// Read testing
