+++ internal/framework/resources/domain_resource_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	_ "github.com/terraform-providers/terraform-provider-dnsimple/internal/framework/resources"
+	"github.com/terraform-providers/terraform-provider-dnsimple/internal/framework/test_utils"
 )
 
 func TestAccDomainResource(t *testing.T) {
@@ -17,7 +18,7 @@ func TestAccDomainResource(t *testing.T) {
 	resourceName := "dnsimple_domain.test"
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testAccPreCheck(t) },
+		PreCheck:                 func() { test_utils.TestAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
 		CheckDestroy:             testAccCheckDomainResourceDestroy,
 		Steps: []resource.TestStep{
