+++ internal/framework/resources/lets_encrypt_certificate_resource_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	_ "github.com/terraform-providers/terraform-provider-dnsimple/internal/framework/resources"
+	"github.com/terraform-providers/terraform-provider-dnsimple/internal/framework/test_utils"
 )
 
 func TestAccLetsEncryptCertificateResource(t *testing.T) {
@@ -24,7 +25,7 @@ func TestAccLetsEncryptCertificateResource(t *testing.T) {
 	certSigAlg := os.Getenv("DNSIMPLE_CERTIFICATE_SIGNATURE_ALGORITHM")
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testAccPreCheck(t) },
+		PreCheck:                 func() { test_utils.TestAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
 		CheckDestroy:             testAccCheckLetsEncryptCertificateResourceDestroy,
 		Steps: []resource.TestStep{
