+++ internal/framework/utils/utils.go
@@ -2,10 +2,15 @@ package utils
 
 import (
 	"context"
+	"fmt"
 	"math/rand"
 	"os"
-	"testing"
+	"strings"
 	"time"
+
+	"github.com/dnsimple/dnsimple-go/dnsimple"
+	"github.com/hashicorp/terraform-plugin-framework/diag"
+	"github.com/hashicorp/terraform-plugin-framework/path"
 )
 
 func GetDefaultFromEnv(key, fallback string) string {
@@ -13,23 +18,6 @@ func GetDefaultFromEnv(key, fallback string) string {
 		return value
 	}
 	return fallback
-}
-
-func TestAccPreCheck(t *testing.T) {
-	// You can add code here to run prior to any test case execution, for example assertions
-	// about the appropriate environment variables being set are common to see in a pre-check
-	// function.
-	if v := os.Getenv("DNSIMPLE_TOKEN"); v == "" {
-		t.Fatal("DNSIMPLE_TOKEN must be set for acceptance tests")
-	}
-
-	if v := os.Getenv("DNSIMPLE_ACCOUNT"); v == "" {
-		t.Fatal("DNSIMPLE_ACCOUNT must be set for acceptance tests")
-	}
-
-	if v := os.Getenv("DNSIMPLE_DOMAIN"); v == "" {
-		t.Fatal("DNSIMPLE_DOMAIN must be set for acceptance tests. The domain is used to create and destroy record against.")
-	}
 }
 
 const alphabet = "0123456789abcdefghijklmnopqrstuvwxyz"
@@ -71,5 +59,37 @@ func RetryWithTimeout(ctx context.Context, fn func() error, timeout time.Duratio
 		case <-time.After(delay):
 			continue
 		}
+	}
+}
+
+func AttributeErrorsToDiagnostics(err *dnsimple.ErrorResponse) diag.Diagnostics {
+	diagnostics := diag.Diagnostics{}
+
+	diagnostics.AddError(
+		"API returned an error",
+		err.Message,
+	)
+
+	for field, errors := range err.AttributeErrors {
+		terraformField := TranslateFieldFromAPIToTerraform(field)
+
+		diagnostics.AddAttributeError(
+			path.Root(terraformField),
+			fmt.Sprintf("API returned a Validation Error for: %s", terraformField),
+			strings.Join(errors, ", "),
+		)
+	}
+
+	return diagnostics
+}
+
+func TranslateFieldFromAPIToTerraform(field string) string {
+	switch field {
+	case "record_type":
+		return "type"
+	case "content":
+		return "value"
+	default:
+		return field
 	}
 }
